+++ packages/f-header/src/components/Logo.vue
@@ -21,8 +21,8 @@
 
 <script>
 import {
-    JeLogoNoColourIcon as JeLogo,
-    MlLogoIcon as MlLogo
+    LogoJusteatIcon as JeLogo,
+    LogoMenulogIcon as MlLogo
 } from '@justeat/f-vue-icons';
 
 export default {
@@ -59,7 +59,7 @@ export default {
             return `Go to ${this.companyName} homepage`;
         },
         logoColourModifier () {
-            return (this.headerBackgroundTheme === 'transparent' || this.headerBackgroundTheme === 'red') ? 'c-icon--alt' : '';
+            return (this.headerBackgroundTheme === 'transparent' || this.headerBackgroundTheme === 'orange') ? 'c-icon--alt' : '';
         }
     }
 };
@@ -72,8 +72,7 @@ export default {
         display: flex;
         justify-content: center;
         height: $header-height--narrow;
-        padding-top: 22px;
-
+        padding-top: 12px;
 
         @include theme(ml) {
             padding-top: 8px;
@@ -82,7 +81,7 @@ export default {
         @include media('>=mid') {
             justify-content: left;
             height: $header-height;
-            padding-top: 24px;
+            padding-top: 20px;
 
             @include theme(ml) {
                 padding-top: 16px;
@@ -92,24 +91,24 @@ export default {
 
     .c-logo-img {
         // default logo image height and width (as should be an inline SVG)
-        width: 82px;
-        height: 16px;
+        width: 98px;
+        height: 24px;
+        margin-left: -10.5px; //half of hamburger menu width
 
         @include media('>=mid') {
-            width: 165px;
-            height: 32px;
+            width: 163px;
+            height: 40px;
+            margin-left: 0;
         }
 
         // Menulog logo, as it has multiple fill values built in. We just hide the outline on transparent mode.
         @include theme(ml) {
             width: 120px;
             height: 32px;
-            margin-left: -10.5px; //half of hamburger menu width
 
             @include media('>=mid') {
                 width: 149px;
                 height: 41px;
-                margin-left: 0;
             }
 
             path:first-child {
@@ -120,11 +119,13 @@ export default {
         }
     }
 
-    .c-icon--je g {
+    .c-logo-img g,
+    .c-logo-img path {
         fill: $header-logo-color;
     }
 
-    .c-icon--alt g {
+    .c-icon--alt g,
+    .c-icon--alt path {
         fill: $header-logo-color--alt;
     }
 </style>